# openapi.yaml
openapi: 3.0.3
info:
  title: MonPetitBiz WhatsApp Bot API
  description: |
    A WhatsApp bot API for micro-businesses in Africa to record transactions, 
    track inventory, and generate financial reports through natural language conversations.
    
    ## Features
    - WhatsApp webhook integration
    - Natural language processing for business commands
    - Transaction recording (sales & expenses)
    - Inventory management
    - Financial report generation
    - PDF report delivery via WhatsApp
    - Comprehensive health monitoring
    
    ## Authentication
    Most endpoints require JWT authentication. Use the `/auth/send-otp` and 
    `/auth/verify-otp` endpoints to obtain an access token.
    
  version: 1.0.0
  contact:
    name: MonPetitBiz Support
    email: support@monpetitbiz.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.monpetitbiz.com
    description: Production server
...
paths:
...
  /whatsapp/webhook:
    get:
      tags:
        - WhatsApp
      summary: Webhook verification
      description: Verifies the webhook during WhatsApp setup
      parameters:
        - name: hub.mode
          in: query
          required: true
          schema:
            type: string
            example: "subscribe"
        - name: hub.challenge
          in: query
          required: true
          schema:
            type: string
            example: "1234567890"
        - name: hub.verify_token
          in: query
          required: true
          schema:
            type: string
            example: "your_verify_token"
      responses:
        '200':
          description: Webhook verified successfully
          content:
            text/plain:
              schema:
                type: string
                example: "1234567890"
        '401':
          description: Invalid verify token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - WhatsApp
      summary: Receive WhatsApp messages
      description: Processes incoming WhatsApp messages and executes business logic
      parameters:
        - name: x-hub-signature-256
          in: header
          description: WhatsApp signature for payload verification
          schema:
            type: string
            example: "sha256=abc123..."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookPayload'
            examples:
              text_message:
                summary: Text message
                value:
                  object: "whatsapp_business_account"
                  entry:
                    - id: "entry_id"
                      changes:
                        - value:
                            messaging_product: "whatsapp"
                            metadata:
                              display_phone_number: "1234567890"
                              phone_number_id: "phone_id"
                            messages:
                              - id: "msg_id"
                                from: "221123456789"
                                timestamp: "1640995200"
                                text:
                                  body: "vente 1000"
                                type: "text"
                          field: "messages"
      responses:
        '200':
          description: Message processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Processed 1 messages"
  /whatsapp/test:
    post:
      tags:
        - WhatsApp
      summary: Test message processing (Development only)
      description: Test endpoint for simulating WhatsApp messages in development
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - from
                - message
              properties:
                from:
                  type: string
                  example: "221123456789"
                message:
                  type: string
                  example: "vente pain 1500"
            examples:
              sale:
                summary: Sale command
                value:
                  from: "221123456789"
                  message: "vente pain 1500"
              expense:
                summary: Expense command
                value:
                  from: "221123456789"
                  message: "dépense 500 marchandise"
              stock:
                summary: Stock update
                value:
                  from: "221123456789"
                  message: "stock pain 50"
              balance:
                summary: Balance request
                value:
                  from: "221123456789"
                  message: "bilan jour"
...
  /whatsapp/twilio:
    post:
      tags:
        - WhatsApp
      summary: Receive Twilio WhatsApp messages (Legacy)
      description: |
        Legacy webhook endpoint for Twilio WhatsApp messages. 
        **Deprecated**: Use `/whatsapp/twilio/webhook` instead.
        
        This endpoint is maintained for backward compatibility but may be removed in future versions.
      deprecated: true
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TwilioWebhookPayload'
      responses:
        '200':
          description: Message processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Message processed successfully"
  /whatsapp/test-twilio:
    post:
      tags:
        - WhatsApp
      summary: Test Twilio WhatsApp webhook format (Development only)
      description: Test endpoint that simulates a Twilio WhatsApp webhook for development purposes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - from
                - message
              properties:
                from:
                  type: string
                  example: "+221123456789"
                  description: "Phone number without whatsapp: prefix"
                message:
                  type: string
                  example: "vente pain 1500"
                  description: "Message content"
                profileName:
                  type: string
                  example: "John Doe"
                  description: "WhatsApp profile name (optional)"
            examples:
              test_sale:
                summary: Test sale command
                value:
                  from: "+221123456789"
                  message: "vente pain 1500"
                  profileName: "Test User"
              test_registration:
                summary: Test registration
                value:
                  from: "+221987654321"
                  message: "bonjour"
                  profileName: "New User"
      responses:
        '200':
          description: Test message processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  twilioPayload:
                    type: object
                  result:
                    type: object
      responses:
        '200':
          description: Test message processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  result:
                    type: object

  # Health Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    WebhookPayload:
      type: object
      description: Meta WhatsApp Cloud API webhook payload (legacy)
      properties:
        object:
          type: string
          example: "whatsapp_business_account"
        entry:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              changes:
                type: array
                items:
                  type: object
                  properties:
                    value:
                      type: object
                      properties:
                        messaging_product:
                          type: string
                          example: "whatsapp"
                        metadata:
                          type: object
                          properties:
                            display_phone_number:
                              type: string
                            phone_number_id:
                              type: string
                        messages:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                type: string
                              from:
                                type: string
                              timestamp:
                                type: string
                              text:
                                type: object
                                properties:
                                  body:
                                    type: string
                              type:
                                type: string
                                enum: [text, document, image, audio, video]
                    field:
                      type: string
                      example: "messages"
    TwilioWebhookPayload:
      type: object
      description: Twilio Programmable Messaging webhook payload for WhatsApp messages
      required:
        - MessageSid
        - AccountSid
        - From
        - To
        - Body
        - WaId
      properties:
        MessageSid:
          type: string
          description: Unique identifier for the message
          example: "SM1234567890abcdef1234567890abcdef"
          pattern: '^SM[a-f0-9]{32}$'
        AccountSid:
          type: string
          description: Twilio Account SID
          example: "AC1234567890abcdef1234567890abcdef"
          pattern: '^AC[a-f0-9]{32}$'
        From:
          type: string
          description: Sender's WhatsApp number in Twilio format
          example: "whatsapp:+221123456789"
          pattern: '^whatsapp:\+[1-9]\d{1,14}$'
        To:
          type: string
          description: Recipient's WhatsApp number in Twilio format
          example: "whatsapp:+14155238886"
          pattern: '^whatsapp:\+[1-9]\d{1,14}$'
        Body:
          type: string
          description: Message content
          example: "vente pain 1500"
          maxLength: 1600
        ProfileName:
          type: string
          description: WhatsApp profile name of the sender (optional)
          example: "Fabrice Ilboudo"
          maxLength: 256
        WaId:
          type: string
          description: WhatsApp ID (phone number without country code formatting)
          example: "221123456789"
          pattern: '^\d{10,15}$'
        SmsStatus:
          type: string
          description: Message status from Twilio
          enum: [received, sent, delivered, failed, undelivered]
          example: "received"
        NumMedia:
          type: string
          description: Number of media attachments (optional)
          example: "0"
          pattern: '^\d+$'
        MediaUrl0:
          type: string
          description: URL of first media attachment (if NumMedia > 0)
          format: uri
          example: "https://api.twilio.com/2010-04-01/Accounts/.../Messages/.../Media/..."
        MediaContentType0:
          type: string
          description: Content type of first media attachment
          example: "image/jpeg"
        MessageStatus:
          type: string
          description: Current status of the message
          enum: [accepted, queued, sending, sent, failed, delivered, undelivered, receiving, received]
          example: "received"
        ErrorCode:
          type: string
          description: Error code if message failed (optional)
          example: "30008"
        ErrorMessage:
          type: string
          description: Error message if message failed (optional)
          example: "Unknown error"
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: number
          example: 400
        message:
          type: string
          example: "Invalid request"
        error:
          type: string
          example: "Bad Request"
        timestamp:
          type: string
          format: date-time
...